third_party/blink/web_tests/http/tests/images/feature-policy-unoptimized-lossless-images-strict-expected.txt

CONSOLE ERROR: Feature policy violation: unoptimized-lossless-images-strict is not allowed in this document.
CONSOLE ERROR: Feature policy violation: unoptimized-lossless-images-strict is not allowed in this document.
